src/main/java/com/gitlab/emradbuba/learning/java_examples/leetcode/likedlistcycle/LinkedListCycle.java
package com.gitlab.emradbuba.learning.java_examples.leetcode.likedlistcycle;

public class LinkedListCycle {
    public boolean hasCycle(ListNode head) {
        if (head == null || head.next == null) return false;
        if (head.next == head) return true;

        ListNode prev = null;
        ListNode curr = head;
        ListNode next;

        while (curr != null) {
            next = curr.next;
            curr.next = prev;
            prev = curr;
            curr = next;
        }
        return prev == head;
    }
}
